Move execution under numpy dir
