+++ src/jinx/execution/README.md
@@ -1,0 +1,3 @@
+# Execution
+
+Jinx supports the execution of J expressions using different backends.
