+++ src/jinx/execution/numpy/conjunctions.py
@@ -7,14 +7,14 @@ import numpy as np
 
 from jinx.vocabulary import Verb, Noun, Monad, Dyad
 from jinx.errors import DomainError, JinxNotImplementedError
-from jinx.execution.application import (
+from jinx.execution.numpy.application import (
     _apply_dyad,
     _apply_monad,
     split_into_cells,
     get_rank,
     fill_and_assemble,
 )
-from jinx.execution.helpers import (
+from jinx.execution.numpy.helpers import (
     maybe_pad_with_fill_value,
     maybe_parenthesise_verb_spelling,
 )
