--- src/jinx/execution/primitives.py
@@ -1,8 +1,0 @@
-"""J primitives."""
-
-from jinx.execution.adverbs import ADVERB_MAP
-from jinx.execution.conjunctions import CONJUNCTION_MAP
-from jinx.execution.verbs import VERB_MAP
-
-
-PRIMITIVE_MAP = ADVERB_MAP | CONJUNCTION_MAP | VERB_MAP
