+++ tests/execution/numpy/test_application.py
@@ -2,13 +2,13 @@ import numpy as np
 import pytest
 
 from jinx.errors import LengthError
-from src.jinx.execution.application import apply_dyad
-from src.jinx.vocabulary import Noun, DataType
-from src.jinx.execution.primitives import (
+from jinx.execution.numpy.application import apply_dyad
+from jinx.vocabulary import Noun, DataType
+from jinx.execution.numpy.primitives import (
     PRIMITIVE_MAP as PRIMITIVE_MAP_NP,
 )
-from src.jinx.execution.conjunctions import rank_conjunction
-from src.jinx.primitives import PRIMITIVE_MAP
+from jinx.execution.numpy.conjunctions import rank_conjunction
+from jinx.primitives import PRIMITIVE_MAP
 
 
 PLUS = PRIMITIVE_MAP["PLUS"]
